Introduce test-common module.
Test common module includes the TestFramework definitions. This is moved
into a common module to be used in multiple integration test files.

A new feature flag `integration-test` is added which enables the test common
and the integration tests. This is required to modify configuration parameters
differently from production builds.

CI updated with new feature flag.
