+++ .github/workflows/coverage.yaml
@@ -25,7 +25,7 @@ jobs:
           CARGO_INCREMENTAL: '0'
           RUSTFLAGS: '-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off'
           RUSTDOCFLAGS: '-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off'
-        run: cargo test
+        run: cargo test --features integration-test
       - id: coverage
         name: Generate coverage
         uses: actions-rs/grcov@v0.1.5
