+++ Cargo.toml
@@ -24,5 +24,7 @@ bitcoind = {version = "0.32", features = ["25_0"] }
 
 #Empty default feature set, (helpful to generalise in github actions)
 [features]
-default = [] 
+default = []
+# Only used for running the integration tests
+integration-test = []
