+++ tests/test_standard_coinswap.rs
@@ -1,8 +1,8 @@
-use bip39::Mnemonic;
-
+#![cfg(feature = "integration-test")]
+use bitcoin::Amount;
 use teleport::{
     maker::MakerBehavior,
-    utill::str_to_bitcoin_network,
+    test_commons::*,
     wallet::{fidelity::YearAndMonth, RPCConfig, Wallet, WalletMode},
 };
 
@@ -15,109 +15,6 @@ use std::{
 };
 
 use std::str::FromStr;
-
-static TEMP_FILES_DIR: &str = "tests/temp-files";
-static TAKER: &str = "tests/temp-files/taker-wallet";
-static MAKER1: &str = "tests/temp-files/maker-wallet-1";
-static MAKER2: &str = "tests/temp-files/maker-wallet-2";
-
-// Helper function to create new wallet
-fn create_wallet_and_import(filename: PathBuf, rpc_config: &RPCConfig) -> Wallet {
-    if filename.exists() {
-        fs::remove_file(&filename).unwrap();
-    }
-    let mnemonic = Mnemonic::generate(12).unwrap();
-    let seedphrase = mnemonic.to_string();
-
-    let mut wallet = Wallet::init(
-        &filename,
-        rpc_config,
-        seedphrase,
-        "".to_string(),
-        Some(WalletMode::Testing),
-    )
-    .unwrap();
-
-    wallet.sync().unwrap();
-
-    wallet
-}
-
-use bitcoin::{Address, Amount};
-
-use bitcoind::{
-    bitcoincore_rpc::{Auth, RpcApi},
-    BitcoinD, Conf,
-};
-
-struct TestFrameWork {
-    bitcoind: BitcoinD,
-}
-
-impl TestFrameWork {
-    pub fn new(conf: Option<Conf>) -> Self {
-        let exe_path = bitcoind::downloaded_exe_path().unwrap();
-        log::debug!("Launching Bitcoind {}", exe_path);
-        let mut conf = conf.unwrap_or_default();
-        conf.args.push("-txindex=1");
-        let bitcoind = BitcoinD::from_downloaded_with_conf(&conf).unwrap();
-        // Generate initial fund
-        let mining_address = bitcoind
-            .client
-            .get_new_address(None, None)
-            .unwrap()
-            .require_network(bitcoind::bitcoincore_rpc::bitcoin::Network::Regtest)
-            .unwrap();
-        bitcoind
-            .client
-            .generate_to_address(101, &mining_address)
-            .unwrap();
-        Self { bitcoind }
-    }
-
-    pub fn generate_1_block(&self) {
-        let mining_address = self
-            .bitcoind
-            .client
-            .get_new_address(None, None)
-            .unwrap()
-            .require_network(bitcoind::bitcoincore_rpc::bitcoin::Network::Regtest)
-            .unwrap();
-        self.bitcoind
-            .client
-            .generate_to_address(1, &mining_address)
-            .unwrap();
-    }
-
-    pub fn send_to_address(&self, addrs: &Address, amount: Amount) {
-        self.bitcoind
-            .client
-            .send_to_address(addrs, amount, None, None, None, None, None, None)
-            .unwrap();
-    }
-}
-
-impl From<&TestFrameWork> for RPCConfig {
-    fn from(value: &TestFrameWork) -> Self {
-        let url = value.bitcoind.rpc_url().split_at(7).1.to_string();
-        let auth = Auth::CookieFile(value.bitcoind.params.cookie_file.clone());
-        let network = str_to_bitcoin_network(
-            value
-                .bitcoind
-                .client
-                .get_blockchain_info()
-                .unwrap()
-                .chain
-                .as_str(),
-        );
-        Self {
-            url,
-            auth,
-            network,
-            ..Default::default()
-        }
-    }
-}
 
 #[tokio::test]
 async fn test_standard_coinswap() {
@@ -256,6 +153,7 @@ async fn test_standard_coinswap() {
             500000,
             2,
             3,
+            None,
         );
     });
 
@@ -265,9 +163,9 @@ async fn test_standard_coinswap() {
         while !*kill_block_creation_clone.read().unwrap() {
             thread::sleep(Duration::from_secs(5));
             test_frameowrk_ptr.generate_1_block();
-            println!("created block");
+            log::info!("created block");
         }
-        println!("ending block creation thread");
+        log::info!("ending block creation thread");
     });
 
     taker_thread.join().unwrap();
@@ -324,4 +222,6 @@ async fn test_standard_coinswap() {
 
     // Remove test temp files and dir
     fs::remove_dir_all::<PathBuf>(TEMP_FILES_DIR.into()).unwrap();
+
+    test_framework.stop();
 }
